feat: enhance pre-update password flow with registration support and token handling
